% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_date.R
\name{add_date}
\alias{add_date}
\title{Turn created_at field into a date variable}
\usage{
add_date(df)
}
\arguments{
\item{df}{A parsed Tweet JSON file (data.frame) which includes created_at variable}
}
\value{
date column is added to the data.frame
}
\description{
Turn created_at field into a date variable
}
