% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_US_location.R
\name{add_US_location}
\alias{add_US_location}
\title{Identity US locations based on location field}
\usage{
add_US_location(df)
}
\arguments{
\item{df}{A parsed Tweet JSON file (data.frame) which includes location variable}
}
\value{
US_location column is added to the data.frame
}
\description{
Identity US locations based on location field
}
