% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_to_dataframe.R
\name{html_to_dataframe}
\alias{html_to_dataframe}
\title{Parse an Ethnic NewsWatch search result (saved in HTML format) into a dataframe}
\usage{
html_to_dataframe(file_path)
}
\arguments{
\item{file_path}{A file path which indicates an HTML file that contains the search results from the Ethnic NewsWatch database. This input should be a string vector.}
}
\value{
A dataframe with four columns ("text", "source", "author", "date")
}
\description{
Parse an Ethnic NewsWatch search result (saved in HTML format) into a dataframe
}
