% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_to_dataframe_all.R
\name{html_to_dataframe_all}
\alias{html_to_dataframe_all}
\title{Parse multiple ProQuest's archive search result (saved in HTML format) into a dataframe simultaneously}
\usage{
html_to_dataframe_all(filepath)
}
\arguments{
\item{filepath}{A file path where a user saved HTML files containing the search results from the ProQuest's news archive. This input should be a string vector.}
}
\value{
A dataframe with four columns ("text", "source", "author", "date")
}
\description{
Parse multiple ProQuest's archive search result (saved in HTML format) into a dataframe simultaneously
}
