% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter_jsonl2_df.R
\name{twitter_jsonl2_df}
\alias{twitter_jsonl2_df}
\title{Parse a Tweet JSON file into a dataframe}
\usage{
twitter_jsonl2_df(file_path)
}
\arguments{
\item{file_path}{A file path which indicates a Tweet JSON file. This input should be a string vector.}
}
\value{
A dataframe with eight columns ("document.id", "ccode", "created_at", "full_text",
																			 "retweet_count", "favorite_count", "user.followers_count", "user.friends_count")
}
\description{
Parse a Tweet JSON file into a dataframe
}
