% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter_jsonl2_df_all.R
\name{twitter_jsonl2_df_all}
\alias{twitter_jsonl2_df_all}
\title{Parse multiple Tweet JSON files into a dataframe}
\usage{
twitter_jsonl2_df_all(dir_path)
}
\arguments{
\item{dir_path}{A directory path where a user saved Tweet JSON files. This input should be a string vector.}
}
\value{
A dataframe with eight columns ("document.id", "ccode", "created_at", "full_text",
																			 "retweet_count", "favorite_count", "user.followers_count", "user.friends_count")
}
\description{
Parse multiple Tweet JSON files into a dataframe
}
