% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_to_dataframe_all.R
\name{html_to_dataframe_all}
\alias{html_to_dataframe_all}
\title{Parse multiple Ethnic NewsWatch search result (saved in HTML format) into a dataframe}
\usage{
html_to_dataframe_all(dir_path)
}
\arguments{
\item{dir_path}{A directory path where a user saved the HTML files containing the search results from the Ethnic NewsWatch database. This input should be a string vector.}
}
\value{
A dataframe with four columns ("text", "source", "author", "date")
}
\description{
Parse multiple Ethnic NewsWatch search result (saved in HTML format) into a dataframe
}
